Prepare for the new login workflow
This sets up the GenerateToken function for the new login workflow.
In order to use it, you'll have to specify the AccountID and UserID
on the Client.

In addition, the "deep" arguments have all been removed.  If you want
to recurse into structures and make additional requests, you're more
than welcome to, but it won't happen automatically.

Finally, the command-line tool now spits out JSON for better
compatibility with other tools.